+++ docs/changelog.rst
@@ -17,6 +17,7 @@ Changelog
 * :bug:`11782` Adding a new EVM account will no longer trigger a redundant initial balance fetch before token detection. Balances are now fetched only once after tokens are detected.
 * :feature:`-` Claiming Degen airdrop 1 will now be properly decoded by rotki.
 * :bug:`11788` Removing a sub-event from a multi-trade event group will no longer fail with a sequence index conflict error.
+* :bug:`11885` Blockscout will not be used to query internal transactions in OP for transactions before the bedrock upgrade.
 * :bug:`-` Claiming Pendle rewards from old pools, or claiming multiple rewards in one transaction should now be decoded properly by rotki.
 * :bug:`-` ETH staking historyMEV/proposer payout rewards are combined more reliably with block proposal even when beaconcha.in relay data is unavailable.
 * :bug:`11766` Value distribution graphs will now work correctly by fixing the schema types to match the backend response.